Improving precision to avoid rounding issues

+++ geometry.c
@@ -7,7 +7,7 @@ point_t point_minus(point_t a, point_t b)
 	return (point_t){a.x-b.x,a.y-b.y};
 }
 
-float point_cross(point_t a, point_t b)
+double point_cross(point_t a, point_t b)
 {
 	return a.x*b.y - a.y*b.x;
 }
@@ -57,7 +57,7 @@ delta = b^2 - 4ac
 
 We have: y = (-b +- sqrt(delta)) / (2a)
 */
-char parabola_intersect(point_t* dst, const point_t* f1, const point_t* f2, float p)
+char parabola_intersect(point_t* dst, const point_t* f1, const point_t* f2, double p)
 {
 	const point_t* f = f1;
 
@@ -72,21 +72,21 @@ char parabola_intersect(point_t* dst, const point_t* f1, const point_t* f2, floa
 	}
 	else
 	{
-		float r = (f2->x-p)/(f1->x-p);
-		float d1 = f1->x*f1->x - p*p + f1->y*f1->y;
-		float d2 = f2->x*f2->x - p*p + f2->y*f2->y;
+		double r = (f2->x-p)/(f1->x-p);
+		double d1 = f1->x*f1->x - p*p + f1->y*f1->y;
+		double d2 = f2->x*f2->x - p*p + f2->y*f2->y;
 
-		float a = r-1;
-		float b = 2*(f2->y - f1->y*r);
-		float c = d1*r - d2;
-		float delta = b*b - 4*a*c;
+		double a = r-1;
+		double b = 2*(f2->y - f1->y*r);
+		double c = d1*r - d2;
+		double delta = b*b - 4*a*c;
 
 		if (delta < 0)
 			return 0;
 		dst->y = (-b + sqrt(delta)) / (2*a);
 	}
 
-	float t = dst->y-f->y;
+	double t = dst->y-f->y;
 	dst->x = (f->x*f->x - p*p + t*t) / (2*(f->x-p));
 	return 1;
 }
@@ -107,16 +107,16 @@ b = (p3.y-p1.y) / (p3.x-p1.x)
 x = (ab(p1.y-p3.y) + b(p1.x+p2.x) - a(p2.x+p3.x)) / (2*(b-a))
 y = -(x - (p1.x+p2.x)/2)/a + (p1.y+p2.y)/2
 
-	float a = (p2->y-p1->y) / (p2->x-p1->x);
-	float b = (p3->y-p2->y) / (p3->x-p2->x);
+	double a = (p2->y-p1->y) / (p2->x-p1->x);
+	double b = (p3->y-p2->y) / (p3->x-p2->x);
 	if (a == b)
 		return 0;
 
-	float x = (a*b*(p1->y-p3->y) + b*(p1->x+p2->x) - a*(p2->x+p3->x)) / (2*(b-a));
-	float y = -(x - (p1->x+p2->x)/2)/a + (p1->y+p2->y)/2;
+	double x = (a*b*(p1->y-p3->y) + b*(p1->x+p2->x) - a*(p2->x+p3->x)) / (2*(b-a));
+	double y = -(x - (p1->x+p2->x)/2)/a + (p1->y+p2->y)/2;
 
-	float dx = p1->x - x;
-	float dy = p1->y - y;
+	double dx = p1->x - x;
+	double dy = p1->y - y;
 	*r = sqrt(dx*dx + dy*dy);
 
 	c->x = x;
@@ -127,7 +127,7 @@ not working
 */
 
 // from http://www.cs.hmc.edu/~mbrubeck/voronoi.html
-char circle_from3(point_t* c, float* r, const point_t* p1, const point_t* p2, const point_t* p3)
+char circle_from3(point_t* c, double* r, const point_t* p1, const point_t* p2, const point_t* p3)
 {
 	// Check that bc is a "right turn" from ap2->
 	if ((p2->x-p1->x)*(p3->y-p1->y) - (p3->x-p1->x)*(p2->y-p1->y) > 0)
@@ -159,12 +159,12 @@ char segment_intersect(point_t* dst, const segment_t* a, const segment_t* b)
 	point_t q = b->a;
 	point_t s = point_minus(b->b,q);
 
-	float y = point_cross(r, s);
+	double y = point_cross(r, s);
 	if (y != 0)
 	{
 		point_t m = point_minus(q, p);
-		float t = point_cross(m, s) / y;
-		float u = point_cross(m, r) / y;
+		double t = point_cross(m, s) / y;
+		double u = point_cross(m, r) / y;
 		if (0 <= t && t <= 1 && 0 <= u && u <= 1)
 		{
 			dst->x = p.x + t*r.x;
